cambios finales parte 2
