--- flutter_application_laboratorio_3/lib/lib/pages/home_page.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MyHomePage extends StatelessWidget {
-  const MyHomePage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: Text("Página Principal")),
-      body: Center(child: Text("¡Bienvenido a Flutter!")),
-    );
-  }
-}
